// src/types/yandex/notification-types.ts
export interface SendNotificationResponse {
    /**
     * Версия интеграции.
     * @minLength 1
     * @maxLength 100
     */
    version: string
    /**
     * Название интеграции.
     * @minLength 1
     * @maxLength 100
     */
    name: string
    /**
     * Дата и время начала обработки уведомления в формате UTC.
     * @format date-time
     */
    time: string
}

/**
 * Обертка для ошибок при обработке уведомления.
 * @example {"error":{"type":"UNKNOWN","message":"message"}}
 */
export interface SendNotificationErrorResponse {
    /** Ошибка при обработке уведомления. */
    error?: NotificationApiErrorDTO
}
// ...
export interface OrderCreatedNotificationDTO {
    /**
     * Тип уведомления:
     *
     * * `PING` — проверочное уведомление.
     * * `ORDER_CREATED` — создан новый заказ.
     * * `ORDER_CANCELLED` — заказ отменен.
     * * `ORDER_STATUS_UPDATED` — статус заказа изменен.
     * * `ORDER_RETURN_CREATED` — создан новый невыкуп или возврат.
     * * `ORDER_CANCELLATION_REQUEST` — создана заявка на отмену заказа (для DBS-магазинов).
     * * `ORDER_RETURN_STATUS_UPDATED` — статус невыкупа или возврата изменен.
     * * `GOODS_FEEDBACK_CREATED` — создан новый отзыв о товаре.
     * * `GOODS_FEEDBACK_COMMENT_CREATED` — создан новый комментарий к отзыву о товаре.
     * * `CHAT_CREATED` — создан новый чат с покупателем.
     * * `CHAT_MESSAGE_SENT` — добавлено новое сообщение в чате.
     * * `CHAT_ARBITRAGE_STARTED` — по обращению покупателя начался спор.
     * * `CHAT_ARBITRAGE_FINISHED` — спор завершен.
     */
    notificationType: NotificationType
    /**
     * Идентификатор заказа.
     * @format int64
     */
    orderId: number
    /**
     * Идентификатор магазина.
     * @format int64
     */
    campaignId: number
    /** Список товаров в заказе. */
    items: NotificationOrderItemDTO[]
    /**
     * Дата и время создания заказа.
     *
     * Формат даты: ISO 8601 со смещением относительно UTC. Например, `2017-11-21T00:00:00.213Z`.
     * @format date-time
     */
    createdAt: string
}
// ...
export interface OrderCancelledNotificationDTO {
    /**
     * Тип уведомления:
     *
     * * `PING` — проверочное уведомление.
     * * `ORDER_CREATED` — создан новый заказ.
     * * `ORDER_CANCELLED` — заказ отменен.
     * * `ORDER_STATUS_UPDATED` — статус заказа изменен.
     * * `ORDER_RETURN_CREATED` — создан новый невыкуп или возврат.
     * * `ORDER_CANCELLATION_REQUEST` — создана заявка на отмену заказа (для DBS-магазинов).
     * * `ORDER_RETURN_STATUS_UPDATED` — статус невыкупа или возврата изменен.
     * * `GOODS_FEEDBACK_CREATED` — создан новый отзыв о товаре.
     * * `GOODS_FEEDBACK_COMMENT_CREATED` — создан новый комментарий к отзыву о товаре.
     * * `CHAT_CREATED` — создан новый чат с покупателем.
     * * `CHAT_MESSAGE_SENT` — добавлено новое сообщение в чате.
     * * `CHAT_ARBITRAGE_STARTED` — по обращению покупателя начался спор.
     * * `CHAT_ARBITRAGE_FINISHED` — спор завершен.
     */
    notificationType: NotificationType
    /**
     * Идентификатор заказа.
     * @format int64
     */
    orderId: number
    /**
     * Идентификатор магазина.
     * @format int64
     */
    campaignId: number
    /** Список товаров в заказе. */
    items: NotificationOrderItemDTO[]
    /**
     * Дата и время отмены заказа.
     *
     * Формат даты: ISO 8601 со смещением относительно UTC. Например, `2017-11-21T00:00:00.213Z`.
     * @format date-time
     */
    cancelledAt: string
}
// ...
export interface OrderReturnCreatedNotificationDTO {
    /**
     * Тип уведомления:
     *
     * * `PING` — проверочное уведомление.
     * * `ORDER_CREATED` — создан новый заказ.
     * * `ORDER_CANCELLED` — заказ отменен.
     * * `ORDER_STATUS_UPDATED` — статус заказа изменен.
     * * `ORDER_RETURN_CREATED` — создан новый невыкуп или возврат.
     * * `ORDER_CANCELLATION_REQUEST` — создана заявка на отмену заказа (для DBS-магазинов).
     * * `ORDER_RETURN_STATUS_UPDATED` — статус невыкупа или возврата изменен.
     * * `GOODS_FEEDBACK_CREATED` — создан новый отзыв о товаре.
     * * `GOODS_FEEDBACK_COMMENT_CREATED` — создан новый комментарий к отзыву о товаре.
     * * `CHAT_CREATED` — создан новый чат с покупателем.
     * * `CHAT_MESSAGE_SENT` — добавлено новое сообщение в чате.
     * * `CHAT_ARBITRAGE_STARTED` — по обращению покупателя начался спор.
     * * `CHAT_ARBITRAGE_FINISHED` — спор завершен.
     */
    notificationType: NotificationType
    /**
     * Идентификатор заказа.
     * @format int64
     */
    orderId: number
    /**
     * Идентификатор невыкупа или возврата.
     * @format int64
     */
    returnId: number
    /**
     * Тип заказа для фильтрации:
     *
     * * `UNREDEEMED` — невыкуп.
     *
     * * `RETURN` — возврат.
     *
     * Если не указывать, в ответе будут и невыкупы, и возвраты.
     */
    returnType: ReturnType
    /**
     * Идентификатор магазина.
     * @format int64
     */
    campaignId: number
    /** Список товаров в невыкупе или возврате. */
    items: NotificationReturnItemDTO[]
    /**
     * Дата и время создания невыкупа или возврата.
     *
     * Формат даты: ISO 8601 со смещением относительно UTC. Например, `2017-11-21T00:00:00.213Z`.
     * @format date-time
     */
    createdAt: string
}

/**
 * Уведомление о смене статуса невыкупа или возврата.
 *
 * `notificationType` = `ORDER_RETURN_STATUS_UPDATED`
 *
 * {% note tip "Чтобы получить подробную информацию о невыкупе или возврате" %}
 *
 * Используйте метод [GET campaigns/{campaignId}/orders/{orderId}/returns/{returnId}](../../reference/orders/getReturn.md).
 *
 * {% endnote %}
 */
export interface OrderReturnStatusUpdatedNotificationDTO {
    /**
     * Тип уведомления:
     *
     * * `PING` — проверочное уведомление.
     * * `ORDER_CREATED` — создан новый заказ.
     * * `ORDER_CANCELLED` — заказ отменен.
     * * `ORDER_STATUS_UPDATED` — статус заказа изменен.
     * * `ORDER_RETURN_CREATED` — создан новый невыкуп или возврат.
     * * `ORDER_CANCELLATION_REQUEST` — создана заявка на отмену заказа (для DBS-магазинов).
     * * `ORDER_RETURN_STATUS_UPDATED` — статус невыкупа или возврата изменен.
     * * `GOODS_FEEDBACK_CREATED` — создан новый отзыв о товаре.
     * * `GOODS_FEEDBACK_COMMENT_CREATED` — создан новый комментарий к отзыву о товаре.
     * * `CHAT_CREATED` — создан новый чат с покупателем.
     * * `CHAT_MESSAGE_SENT` — добавлено новое сообщение в чате.
     * * `CHAT_ARBITRAGE_STARTED` — по обращению покупателя начался спор.
     * * `CHAT_ARBITRAGE_FINISHED` — спор завершен.
     */
    notificationType: NotificationType
    /**
     * Идентификатор заказа.
     * @format int64
     */
    orderId: number
    /**
     * Идентификатор невыкупа или возврата.
     * @format int64
     */
    returnId: number
    /**
     * Идентификатор магазина.
     * @format int64
     */
    campaignId: number
    /**
     * Информация об обновлении статуса невыкупа или возврата.
     *
     * Возвращается только тот статус, который был изменен.
     *
     * Для невыкупов приходит только `shipmentStatus`.
     *
     * Параметр `shipmentStatus` не приходит для возвратов с опцией **Быстрый возврат денег за дешевый брак**, когда товар остается у покупателя.
     */
    statuses: NotificationUpdatedReturnStatusesDTO
    /**
     * Дата и время изменения статуса невыкупа или возврата.
     *
     * Формат даты: ISO 8601 со смещением относительно UTC. Например, `2017-11-21T00:00:00.213Z`.
     * @format date-time
     */
    updatedAt: string
}
// ...
export interface GoodsFeedbackCreatedNotificationDTO {
    /**
     * Тип уведомления:
     *
     * * `PING` — проверочное уведомление.
     * * `ORDER_CREATED` — создан новый заказ.
     * * `ORDER_CANCELLED` — заказ отменен.
     * * `ORDER_STATUS_UPDATED` — статус заказа изменен.
     * * `ORDER_RETURN_CREATED` — создан новый невыкуп или возврат.
     * * `ORDER_CANCELLATION_REQUEST` — создана заявка на отмену заказа (для DBS-магазинов).
     * * `ORDER_RETURN_STATUS_UPDATED` — статус невыкупа или возврата изменен.
     * * `GOODS_FEEDBACK_CREATED` — создан новый отзыв о товаре.
     * * `GOODS_FEEDBACK_COMMENT_CREATED` — создан новый комментарий к отзыву о товаре.
     * * `CHAT_CREATED` — создан новый чат с покупателем.
     * * `CHAT_MESSAGE_SENT` — добавлено новое сообщение в чате.
     * * `CHAT_ARBITRAGE_STARTED` — по обращению покупателя начался спор.
     * * `CHAT_ARBITRAGE_FINISHED` — спор завершен.
     */
    notificationType: NotificationType
    /**
     * Идентификатор отзыва.
     * @format int64
     */
    feedbackId: number
    /**
     * Идентификатор кабинета.
     * @format int64
     */
    businessId: number
    /**
     * Дата и время создания отзыва.
     *
     * Может отличаться от информации в `publishedAt`, так как некоторое время отзыв проходит модерацию.
     *
     * Формат даты: ISO 8601 со смещением относительно UTC. Например, `2017-11-21T00:00:00.213Z`.
     * @format date-time
     */
    createdAt: string
    /**
     * Дата и время публикации отзыва.
     *
     * Может отличаться от информации в `createdAt`, так как некоторое время отзыв проходит модерацию.
     *
     * Формат даты: ISO 8601 со смещением относительно UTC. Например, `2017-11-21T00:00:00.213Z`.
     * @format date-time
     */
    publishedAt: string
}
// ...
export enum NotificationType {
    PING = 'PING',
    ORDER_CREATED = 'ORDER_CREATED',
    ORDER_CANCELLED = 'ORDER_CANCELLED',
    ORDER_STATUS_UPDATED = 'ORDER_STATUS_UPDATED',
    ORDER_RETURN_CREATED = 'ORDER_RETURN_CREATED',
    ORDER_CANCELLATION_REQUEST = 'ORDER_CANCELLATION_REQUEST',
    ORDER_RETURN_STATUS_UPDATED = 'ORDER_RETURN_STATUS_UPDATED',
    GOODS_FEEDBACK_CREATED = 'GOODS_FEEDBACK_CREATED',
    GOODS_FEEDBACK_COMMENT_CREATED = 'GOODS_FEEDBACK_COMMENT_CREATED',
    CHAT_CREATED = 'CHAT_CREATED',
    CHAT_MESSAGE_SENT = 'CHAT_MESSAGE_SENT',
    CHAT_ARBITRAGE_STARTED = 'CHAT_ARBITRAGE_STARTED',
    CHAT_ARBITRAGE_FINISHED = 'CHAT_ARBITRAGE_FINISHED'
}

/**
 * Ошибка при обработке уведомления.
 * @example {"type":"UNKNOWN","message":"message"}
 */
export interface NotificationApiErrorDTO {
    /**
     * Тип ошибки:
     *
     * * `UNKNOWN` — неизвестная ошибка.
     * * `WRONG_EVENT_FORMAT` — неправильный тип уведомления.
     * * `DUPLICATED_EVENT` — дублирующее уведомление.
     */
    type?: NotificationApiErrorType
    /** Описание ошибки. */
    message?: string
}

/** Информация о товаре в заказе. */
export interface NotificationOrderItemDTO {
    /**
     * Ваш SKU — идентификатор товара в вашей системе.
     *
     * Правила использования SKU:
     *
     * * У каждого товара SKU должен быть свой.
     *
     * * Уже заданный SKU нельзя освободить и использовать заново для другого товара. Каждый товар должен получать новый идентификатор, до того никогда не использовавшийся в вашем каталоге.
     *
     * SKU товара можно изменить в кабинете продавца на Маркете. О том, как это сделать, читайте [в Справке Маркета для продавцов](https://yandex.ru/support2/marketplace/ru/assortment/operations/edit-sku).
     *
     * [Что такое SKU и как его назначать](https://yandex.ru/support/marketplace/assortment/add/index.html#fields)
     * @minLength 1
     * @maxLength 255
     * @pattern ^(?=.*\S.*)[^\x00-\x08\x0A-\x1f\x7f]{1,255}$
     */
    offerId: string
    /** Количество товара. */
    count: number
}
// ...
export enum ReturnType {
    UNREDEEMED = 'UNREDEEMED',
    RETURN = 'RETURN'
}

/** Информация о товаре в невыкупе или возврате. */
export interface NotificationReturnItemDTO {
    /**
     * Ваш SKU — идентификатор товара в вашей системе.
     *
     * Правила использования SKU:
     *
     * * У каждого товара SKU должен быть свой.
     *
     * * Уже заданный SKU нельзя освободить и использовать заново для другого товара. Каждый товар должен получать новый идентификатор, до того никогда не использовавшийся в вашем каталоге.
     *
     * SKU товара можно изменить в кабинете продавца на Маркете. О том, как это сделать, читайте [в Справке Маркета для продавцов](https://yandex.ru/support2/marketplace/ru/assortment/operations/edit-sku).
     *
     * [Что такое SKU и как его назначать](https://yandex.ru/support/marketplace/assortment/add/index.html#fields)
     * @minLength 1
     * @maxLength 255
     * @pattern ^(?=.*\S.*)[^\x00-\x08\x0A-\x1f\x7f]{1,255}$
     */
    offerId: string
    /** Количество товара. */
    count: number
}
// ...
export enum NotificationApiErrorType {
    UNKNOWN = 'UNKNOWN',
    WRONG_EVENT_FORMAT = 'WRONG_EVENT_FORMAT',
    DUPLICATED_EVENT = 'DUPLICATED_EVENT'
}